+++ SopApi/Model/Entidades/SEEquivalence.cs
@@ -8,8 +8,8 @@ namespace SopApi.Model.Entidades
     public class SEEquivalence
     {
         public int IdEquivalence { get; set; }
-        public Int16 IdParameter { get; set; }
-        public Int16 IdEntity { get; set; }
+        public byte IdParameter { get; set; }
+        public byte IdEntity { get; set; }
         public Int16 IdCode { get; set; }
         public string Description { get; set; }
         public bool Active { get; set; }
