+++ SopApi/Model/Entidades/SETypeParameter.cs
@@ -7,7 +7,7 @@ namespace SopApi.Model.Entidades
 {
     public class SETypeParameter
     {
-        public Int16 IdParameter { get; set; }
+        public byte IdParameter { get; set; }
         public string Description { get; set; }
         public bool Active { get; set; }
     }
